refactor: optimize connection module export structure
