+++ packages/connection/src/index.ts
@@ -1,7 +1,1 @@
-export * from './node';
-// export * from './browser/stub-client';
-export * from './common/proxy';
-
-export * from './common/ws-channel';
-export * from './browser/ws-channel-handler';
-export * from './common/rpcProtocol';
+export * from './common';
